refactored api endpoint
