+++ src/service/bot.ts
@@ -1,0 +1,3 @@
+import { Bot } from "grammy";
+
+export const bot = new Bot(process.env.BOT_TOKEN!);
